Refactor: Change distance data type to Float for precision
This commit refactors the data type used for tracking course distance from `Int` to `Float`. This change provides greater precision when calculating and storing the distance traveled by the user, especially for smaller increments.

### Key Changes:

-   **Data Models & Preferences**:
    -   In `CoursePreference.kt` and `UserRouteScope.kt`, the `courseDistance` and `distance` properties have been changed from `Int` to `Float`.
    -   The default and invalid-state values in `CoursePreference` were updated to use floating-point numbers (e.g., `-1.0f`).

-   **`RouteRecordingService.kt`**:
    -   The service now accumulates distance as a `Float`, ensuring that fractional meter values from location updates are not lost due to integer conversion.

-   **`CourseRepository.kt`**:
    -   When retrieving or saving course data, the `distance` is now handled as a `Float`.
    -   For backward compatibility with APIs expecting an integer, the float value is rounded to the nearest integer (`roundToInt()`) only when finalizing the data for submission.

+++ data/src/main/java/com/who_summoned_the_cloud/eromoro/data/model/UserRouteScope.kt
@@ -4,5 +4,5 @@ import com.who_summoned_the_cloud.eromoro.common.model.Position
 
 interface UserRouteScope {
     var userRoute: List<Position>
-    var distance: Int
+    var distance: Float
 }
